+++ LimitOrderBookSimulation/LimitOrderBooks/ILimitOrderBook.cs
@@ -8,9 +8,9 @@ namespace LimitOrderBookSimulation.LimitOrderBooks
 
         double Time { set; get; }
 
-        long Ask { get; }
+        int Ask { get; }
 
-        long Bid { get; }
+        int Bid { get; }
         
         #endregion Properties
 
@@ -20,25 +20,25 @@ namespace LimitOrderBookSimulation.LimitOrderBooks
 
         #region Limit order
 
-        void SubmitLimitBuyOrder(long price, long amount = 1);
+        void SubmitLimitBuyOrder(int price, int amount = 1);
 
-        void SubmitLimitSellOrder(long price, long amount = 1);
+        void SubmitLimitSellOrder(int price, int amount = 1);
 
         #endregion Limit order
 
         #region Market order
 
-        void SubmitMarketBuyOrder(long amount =1);
+        void SubmitMarketBuyOrder(int amount =1);
 
-        void SubmitMarketSellOrder(long amount =1);
+        void SubmitMarketSellOrder(int amount =1);
 
         #endregion Market order
         
         #region Cancel order
 
-        void CancelLimitBuyOrder(long price, long amount = 1);
+        void CancelLimitBuyOrder(int price, int amount = 1);
 
-        void CancelLimitSellOrder(long price, long amount = 1);
+        void CancelLimitSellOrder(int price, int amount = 1);
 
         #endregion Cancel order
 
@@ -46,7 +46,7 @@ namespace LimitOrderBookSimulation.LimitOrderBooks
         
         #region Time evolution
 
-        Dictionary<LimitOrderBookEvent, long> Counter { get; }
+        Dictionary<LimitOrderBookEvent, int> Counter { get; }
 
         /// <summary>
         /// Time-dependent pricing information 
@@ -59,21 +59,21 @@ namespace LimitOrderBookSimulation.LimitOrderBooks
 
         #region Number of orders
 
-        long NumberOfBuyOrders(long minPrice = 0, long maxPrice = long.MaxValue);
+        int NumberOfBuyOrders(int minPrice = 0, int maxPrice = int.MaxValue);
 
-        long NumberOfSellOrders(long minPrice = 0, long maxPrice = long.MaxValue);
+        int NumberOfSellOrders(int minPrice = 0, int maxPrice = int.MaxValue);
 
-        long NumberOfLimitOrders(long minPrice = 0, long maxPrice = long.MaxValue);
+        int NumberOfLimitOrders(int minPrice = 0, int maxPrice = int.MaxValue);
 
         #endregion Number of orders
 
         #region Inverse CDF
 
-        long InverseCDF(long minPrice, long maxPrice, long q);
+        int InverseCDF(int minPrice, int maxPrice, int q);
 
-        long InverseCDFSellSide(long minPrice, long maxPrice, long q);
+        int InverseCDFSellSide(int minPrice, int maxPrice, int q);
 
-        long InverseCDFBuySide(long minPrice, long maxPrice, long q);
+        int InverseCDFBuySide(int minPrice, int maxPrice, int q);
 
         #endregion Inverse CDF
         
@@ -81,8 +81,8 @@ namespace LimitOrderBookSimulation.LimitOrderBooks
         
         #region Iinitialize
 
-        void InitializeDepthProfileBuySide(IDictionary<long, long> depthProdile);
-        void InitializeDepthProfileSellSide(IDictionary<long, long> depthProdile);
+        void InitializeDepthProfileBuySide(IDictionary<int, int> depthProdile);
+        void InitializeDepthProfileSellSide(IDictionary<int, int> depthProdile);
         
         #endregion Iinitialize
